+++ src/NEventStore.Persistence.PostgreSql.Tests/PersistenceEngineFixture.cs
@@ -1,8 +1,6 @@
 ﻿using NEventStore.Persistence.Sql.Tests;
 using NEventStore.Persistence.Sql;
 using NEventStore.Persistence.Sql.SqlDialects;
-using NEventStore.Serialization;
-using System;
 using NEventStore.Serialization.Binary;
 
 namespace NEventStore.Persistence.AcceptanceTests
@@ -11,6 +9,9 @@ namespace NEventStore.Persistence.AcceptanceTests
 	{
 		public PersistenceEngineFixture()
 		{
+#if NET8_0_OR_GREATER
+			AppContext.SetSwitch("System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization", true);
+#endif
 			// It will be done when creating the PostgreNpgsql6Dialect dialect
 			// AppContext.SetSwitch("Npgsql.EnableLegacyTimestampBehavior", true);
 
@@ -30,7 +31,7 @@ namespace NEventStore.Persistence.AcceptanceTests
 			{
 				var serializer = new BinarySerializer();
 				return new SqlPersistenceFactory(
-					new EnviromentConnectionFactory("PostgreSql", Npgsql.NpgsqlFactory.Instance),
+					new EnvironmentConnectionFactory("PostgreSql", Npgsql.NpgsqlFactory.Instance),
 					serializer,
 					new DefaultEventSerializer(serializer),
 					new PostgreNpgsql6Dialect(npgsql6timestamp: true),