+++ src/NEventStore.Persistence.Sql/ConnectionScope.cs
@@ -4,6 +4,7 @@ namespace NEventStore.Persistence.Sql
 {
 	/// <summary>
 	/// Represents a connection scope.
+	/// Creating and managing a connection can be expensive, so we want to make sure we only create one connection per thread.
 	/// Wraps a DbConnection so that we create no more than one connection for each thread.
 	/// It's a Disposable class, so all the opened connection are closed when the scope is disposed.
 	/// </summary>
@@ -15,14 +16,5 @@ namespace NEventStore.Persistence.Sql
 		public ConnectionScope(string connectionName, Func<DbConnection> factory)
 			: base(connectionName, factory)
 		{ }
-
-		/// <summary>
-		/// Implicitly converts a connection scope to a connection.
-		/// Unwraps the connection from the scope.
-		/// </summary>
-		public static implicit operator DbConnection(ConnectionScope scope)
-		{
-			return scope.Current;
-		}
 	}
 }