+++ Changelog.md
@@ -1,5 +1,13 @@
 # NEventStore.Persistence.Sql
 
+## vNext
+
+- async methods
+
+### Breaking Changes
+
+- `ISerializeEvents.SerializeEventMessages()` signature changed to accept `IEnumerable` instead of `IReadOnlyList`.
+
 ## 9.3.1
 
 - Added a dedicated `ISerializeEvents` interface that allows customizing event deserialization with access to metadata from the `ICommit` class. [#47](https://github.com/NEventStore/NEventStore.Persistence.SQL/issues/47), [#49](https://github.com/NEventStore/NEventStore.Persistence.SQL/issues/49), [#506](https://github.com/NEventStore/NEventStore/issues/506)